requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.2
six==1.16.0
tzdata==2023.4
urllib3==1.26.15
